+++ ui-test/ui-test.jsx
@@ -20,6 +20,8 @@ import ScrollingMixin from './../src/mixins/ScrollingMixin';
 
 // component
 import {
+    BreadcrumbList,
+    BreadcrumbItem,
     ContextMenu,
     MenuItem,
     Nothing,
@@ -152,6 +154,18 @@ const Page = React.createClass({
                     }
                 </Header>
                 <Content>
+                    <BreadcrumbList className="my-own-class">
+                        <BreadcrumbItem onClick={function() {alert('Click on breadcrumb item.');}}>
+                            Button
+                        </BreadcrumbItem>
+                        <BreadcrumbItem href="#">
+                            Link
+                        </BreadcrumbItem>
+                        <BreadcrumbItem>
+                            Span
+                        </BreadcrumbItem>
+                    </BreadcrumbList>
+                    <hr className="mdl-layout-spacer"/>
                     {testcases}
                 </Content>
                 <footer className="mdl-mini-footer">