+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ### Added
 - 'ScrollingMixin' mixin to scroll elements into the visible viewport of scrollable areas
+- provide `<BreadcrumbList/>` and `<BreadcrumbItem/>` to create breadcrumb navigations
 
 ## [2.18.2] 2017-08-31
 ### Added